Prints  '-' instead of 'None' when data is None
Client prints 'None' when data is None , and client also prints
'None' when data is a string 'None'. But string 'None' is  different
to None. To clear the confusion, Client should print '-' instead of
'None' when data is None. Nova Client and Cinder Client has cleared
the confusion.

Change-Id: I0b1496ee78672cb2d841b6283360ba41a94b72b2
Related-Bug: #1266414
Related-Bug: #1422244
Related-Bug: #1600553